Applications: FontEditor, relocate form items by fonts size
FontEditor widget rewritten for respect to the font size and added fixed width and height header values.

+++ Applications/FontEditor/GlyphEditorWidget.cpp
@@ -39,7 +39,7 @@ GlyphEditorWidget::~GlyphEditorWidget()
 {
 }
 
-void GlyphEditorWidget::set_glyph(u8 glyph)
+void GlyphEditorWidget::set_glyph(int glyph)
 {
     if (m_glyph == glyph)
         return;
